Rubocop: move multi-line assignment to next line
This is a small sample of what it would look like if we enable the
`Layout/MultilineAssignmentLayout` rubocop rule. This rule dictates how
we should align assignment across multiple lines. The default wants us
to indent the block on the next line, like so:

```rb
foo =
  if expression
    'bar'
  end
```

There is a configuration for `same_line` which would instead expect us
to format our code like so:

```rb
foo = if expression
  'bar'
end
```

Most of our code is in line with the second option, `same_line`, though
I kind of prefer the first. Once we agree on an approach I'll switch up
the rest of the code to match it.

https://rubocop.readthedocs.io/en/stable/cops_layout/#layoutmultilineassignmentlayout